Disallow null pointers to diffractometer instruments
All low-level data-readers assume a _valid_ (raw) pointer to an
instance of diffractometer instrument (see eg., `core/raw/IDataReader`
or `core/loader/HDF5MetaDataReader`). Therefore, this logical
requirement must be imposed at a high level to prevent undefined
behaviour.

This validity condition is currently verified in the constructor of
`core/experiment/DataHandler` and an exception is thrown in case of an
invalid instrument.

The exception is caught in the GUI session, `gui/models/Session::loadData`.

Resolves issue [224](https://jugit.fz-juelich.de/mlz/nsxtool/-/issues/224)