Eliminate bugs in initializing an Experiment
The bugs caused by commit (48ffc8e9) prevented loading Experiments.

The check for the validity of the pointer to an instance of
diffractometer is moved to `Session::loadData`, so that `Experiment` can
be instantiated with an empty (null) diffractometer.

It is still possible to cause undefined behaviour by accessing a null
diffractometer pointer from `Experiment`, esp. in low-level data-readers.

The error message is slightly improved.