Store and display the names of the source files
The source files used to make a DataSet are stored as a list of strings in `DataSet.sources` and as a string in its metadata.
DataSet has its own MetaData, rather than using its reader's MetaData which is volatile.
In `Session.loadRawData`, `DataSet.MetaData` is accumulated from the individual metadata produced by the reader.

`NumorProperty::onChanged` displays the string-typed contents DataSet metadata as a long scrollable text using `QTextEdit`, so that the sources list be readable.

Code slightly improved.

Resolves issue #87, #89

TODO: The long metadata strings should be stored correctly in the HDF5 file (`ExperimentExporter.writeData`, section for 'Info'). The length is currently limited to 80 chars.