Decouple the name of the DataSet and its associated filename(s)
A confusion between the name of the DataSet and its associated
filename(s) had caused a lot of obscure bugs and crashes.

Currently the name of the dataset is decided by the user.
The filenames are stored and retrieved internally, only if needed.
Several unneeded and confusing methods and attributes (related to
filenames) are removed.