Implement ValueTupleItem to hold and sort table items which include a stdev. or quality measure
`ValueTupleItem` is a `QStandardItem` which holds a floating-point value and its
standard deviation or quality measure.
The sorting mechanism is taylored to sort according the value and then (if the
values are equal) the std. devs. or quality measures.
Sorting is based on the stored _numerical_ values, _not_ the textual representation.

* Sorting examples:
ValueTupleItem("1.2(3)", 1.238, 3e-3) < ValueTupleItem("1.0(3)", 1.01, 3e-3)
ValueTupleItem("1.2(3)", 1.238, 3e-3) < ValueTupleItem("1.2(3)", 1.240, 3e-3)
ValueTupleItem("1.2(3)", 1.238, 3e-3) < ValueTupleItem("1.2(4)", 1.238, 4e-3)

See the Qt documentation for `QStandardItem::operator<`.

Resolves issue #90