ci: optimize ci jobs on release
`GIT_STRATEGY: none` allows to not check out the repo when is not needed.

macOS and Windows jobs are executed on standalone runners, and the project folder
is not cleaned up when there is `GIT_STRATEGY: none`. Therefore, no need to pass
any artifacts, enough to specify the dependency.

`upload` and `release` jobs do not need to checkout the repo at all.